MOSTregi/requirements.txt
astroid==1.6.1
Babel==2.5.3
Django==2.0.2
django-bootstrap3==9.1.0
django-extensions==2.0.0
django-jquery==3.1.0
django-phonenumber-field==2.0.0
django-split-json-widget==1.16
isort==4.3.4
lazy-object-proxy==1.3.1
logilab-astng==0.24.3
logilab-common==1.4.1
mccabe==0.6.1
phonenumberslite==8.9.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pylint==1.8.2
pylint-django==0.9.0
pylint-plugin-utils==0.2.6
pytz==2018.3
six==1.11.0
typing==3.6.4
wrapt==1.10.11
